requirements.txt
gunicorn
fastai
https://download.pytorch.org/whl/cpu/torch-1.1.0-cp36-cp36m-linux_x86_64.whl
flask
flask_jsonpify
requests
werkzeug.utils
